requirements.txt
Appium-Python-Client>=2.11.1
selenium>=4.11.2
pytest>=7.4.0
opencv-python>=4.8.0
numpy>=1.24.0
webdriver-manager>=4.0.1 